requirements.txt
django-bootstrap-v5
python-ldap
